Improved code indentation and reformatting.
